fix: smooth chat scroll after sending a message
